+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [fj-doc-freemarker] support for source type in FreemarkerDocProcessConfig <https://github.com/fugerit-org/fj-doc/issues/223>
+
 ## [8.9.5] - 2024-10-17
 
 ### Changed